# contracts/openapi.yaml
...
components:
  schemas:
...
    Body_uploadDocument:
      properties:
        file:
          contentMediaType: application/octet-stream
          description: 待解析文件。白名单：PDF / DOCX / CSV；单文件 ≤ 100MB
          title: File
          type: string
      required:
      - file
      title: Body_uploadDocument
      type: object
...
    DocumentError:
      description: '`documents.status = failed` 时的失败信息。


        与顶层 `ErrorResponse` 的区别：这里**不含** `trace_id`，

        因为响应体已在外层携带同一个 `trace_id`。'
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
          description: 失败错误码。进程重启回收的场景为 `TASK_INTERRUPTED`（ADR-0001 §3.2）
        detail:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          description: 失败明细摘要。**禁止**回显 `documents.error_detail` 原文（敏感）
          title: Detail
        message:
          title: Message
          type: string
      required:
      - code
      - message
      title: DocumentError
      type: object
    DocumentGraphResponse:
      description: '`GET /api/v1/documents/{id}/graph` 响应：该文档在 Neo4j 中的子图。


        **一致性约束（ADR-0002 §3.2）**：只返回 `status = active` 的 `kg_version`；

        该文档无 active 版本时返回 `409 KG_VERSION_NOT_ACTIVE`，**绝不静默降级**到其他版本。


        规模上限对齐 M3 §3 验收 1：单次返回节点数 ≤ 500，超限时 `truncated = true`。'
      example:
        doc_id: 3f1a9c2e-7b45-4d8a-9e01-2c4f6a8b0d11
        edges:
        - id: r-001
          properties:
            share_pct: 51.0
            since: '2021-03-01'
          source: e-001
          target: e-002
          type: AFFILIATED_WITH
        kg_version: 20260320T1430Z-01H9X9ABCDEF
        node_count: 1
        nodes:
        - canonical_name: 示例科技有限公司
          confidence: 0.93
          entity_type: 公司
          id: e-001
          kg_version: 20260320T1430Z-01H9X9ABCDEF
          label: Entity
        relation_count: 1
        trace_id: 5f2c1b7e-9d4a-4c1e-8f3b-6a0d2e5c7b91
        truncated: false
        version_status: active
      properties:
        doc_id:
          format: uuid
          title: Doc Id
          type: string
        edges:
          items:
            $ref: '#/components/schemas/GraphEdge'
          title: Edges
          type: array
        kg_version:
          description: 本次返回的图谱版本，来自 `GET /internal/kg/active` 的 active 版本
          title: Kg Version
          type: string
        node_count:
          description: '`nodes` 实际条数（截断后）'
          minimum: 0.0
          title: Node Count
          type: integer
        nodes:
          items:
            $ref: '#/components/schemas/GraphNode'
          title: Nodes
          type: array
        relation_count:
          description: '`edges` 实际条数（截断后）'
          minimum: 0.0
          title: Relation Count
          type: integer
        trace_id:
          title: Trace Id
          type: string
        truncated:
          default: false
          description: 是否因超过 500 节点上限被截断
          title: Truncated
          type: boolean
        version_status:
          const: active
          description: 恒为 `active`，是 ADR-0002 一致性强制的可见契约
          title: Version Status
          type: string
      required:
      - doc_id
      - kg_version
      - version_status
      - nodes
      - edges
      - node_count
      - relation_count
      - trace_id
      title: DocumentGraphResponse
      type: object
    DocumentListItem:
      description: '`GET /documents` 单条结果。


        字段对齐 `frontend/src/types/mock.d.ts::DocumentListItem`（Sprint 5 批次 C 收口），

        保证 `npm run gen:api` 生成的 TS 类型可直接替换前端 mock 类型。'
      example:
        entity_count: 128
        file_type: PDF
        filename: e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
        id: 3f1a9c2e-7b45-4d8a-9e01-2c4f6a8b0d11
        status: completed
        task_id: 3f1a9c2e-7b45-4d8a-9e01-2c4f6a8b0d11
        time_label: '10:42'
        trace_id: 5f2c1b7e-9d4a-4c1e-8f3b-6a0d2e5c7b91
        uploaded_at: '2026-09-21T10:42:00Z'
      properties:
        entity_count:
          anyOf:
          - minimum: 0.0
            type: integer
          - type: 'null'
          title: Entity Count
        file_type:
          description: 文件类型标签
          enum:
          - PDF
          - DOCX
          - CSV
          title: File Type
          type: string
        filename:
          description: SHA-256(filename) — 文件名展示用
          title: Filename
          type: string
        id:
          description: 文档主键 UUID
          format: uuid
          title: Id
          type: string
        status:
          description: M1 状态机取值
          enum:
          - pending
          - processing
          - completed
          - failed
          title: Status
          type: string
        task_id:
          description: 任务 id，与 `documents.id` 一致
          format: uuid
          title: Task Id
          type: string
        time_label:
          description: 相对时间文案，固定 UTC 计算
          title: Time Label
          type: string
        trace_id:
          description: 贯穿上传→解析→建图全链路的 trace_id
          title: Trace Id
          type: string
        uploaded_at:
          description: ISO8601 上传时间（UTC）
          title: Uploaded At
          type: string
      required:
      - id
      - filename
      - file_type
      - status
      - uploaded_at
      - time_label
      - task_id
      - trace_id
      title: DocumentListItem
      type: object
    DocumentListResponse:
      description: '`GET /documents` 响应。'
      example:
        items:
        - entity_count: 128
          file_type: PDF
          filename: e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
          id: 3f1a9c2e-7b45-4d8a-9e01-2c4f6a8b0d11
          status: completed
          task_id: 3f1a9c2e-7b45-4d8a-9e01-2c4f6a8b0d11
          time_label: '10:42'
          trace_id: 5f2c1b7e-9d4a-4c1e-8f3b-6a0d2e5c7b91
          uploaded_at: '2026-09-21T10:42:00Z'
        page: 1
        page_size: 10
        total: 27
        trace_id: 5f2c1b7e-9d4a-4c1e-8f3b-6a0d2e5c7b91
      properties:
        items:
          description: 当前页结果
          items:
            $ref: '#/components/schemas/DocumentListItem'
          title: Items
          type: array
        page:
          description: 当前页码（1-based）
          minimum: 1.0
          title: Page
          type: integer
        page_size:
          description: 每页条目数（请求参数回显）
          minimum: 1.0
          title: Page Size
          type: integer
        total:
          description: 当前租户下满足过滤条件的文档总数
          minimum: 0.0
          title: Total
          type: integer
        trace_id:
          title: Trace Id
          type: string
      required:
      - total
      - items
      - page
      - page_size
      - trace_id
      title: DocumentListResponse
      type: object
    DocumentStatusResponse:
      description: '`GET /api/v1/documents/{id}/status` 响应（M1 验收 5）。


        前端轮询建议：首次 2s，3 次后降为 10s。'
      example:
        status: processing
        task_id: 3f1a9c2e-7b45-4d8a-9e01-2c4f6a8b0d11
        trace_id: 5f2c1b7e-9d4a-4c1e-8f3b-6a0d2e5c7b91
      properties:
        error:
          anyOf:
          - $ref: '#/components/schemas/DocumentError'
          - type: 'null'
          description: 仅当 `status = failed` 时非空
        progress:
          anyOf:
          - maximum: 1.0
            minimum: 0.0
            type: number
          - type: 'null'
          description: 0–1 的粗粒度进度。解析进度不可精确估计时为 `null`（Sprint 1 恒为 null）
          title: Progress
        status:
          description: 仅取 `pending / processing / completed / failed`（H1）
          enum:
          - pending
          - processing
          - completed
          - failed
          title: Status
          type: string
        task_id:
          format: uuid
          title: Task Id
          type: string
        trace_id:
          title: Trace Id
          type: string
      required:
      - task_id
      - status
      - trace_id
      title: DocumentStatusResponse
      type: object
...
    ErrorCode:
      description: 统一业务错误码。HTTP 状态码与业务错误码分离（CODEBUDDY.md 错误响应规范）。
      enum:
      - VALIDATION_ERROR
      - UNAUTHORIZED
      - FORBIDDEN
      - NOT_FOUND
      - DOCUMENT_NOT_FOUND
      - ENTITY_NOT_FOUND
      - FILE_TOO_LARGE
      - UNSUPPORTED_MEDIA_TYPE
      - KG_VERSION_NOT_ACTIVE
      - KG_TENANT_LEAK
      - TASK_INTERRUPTED
      - NOT_IMPLEMENTED
      - INTERNAL_ERROR
      - HTTP_ERROR
      title: ErrorCode
      type: string
      x-enum-descriptions:
        DOCUMENT_NOT_FOUND: 文档不存在或未在本租户可见范围内。
        ENTITY_NOT_FOUND: 实体不存在或不属于当前 active kg_version；与跨租户 403 区分。
        FILE_TOO_LARGE: 上传文件超过 MAX_UPLOAD_SIZE_MB（默认 100MB，M1 验收 2）。
        FORBIDDEN: 跨租户访问被拒（ADR-0003：org_id 不符），非本租户资源一律拒绝。
        HTTP_ERROR: 未在错误码表中登记的 HTTP 状态兜底，保留原始 HTTP 状态语义。
        INTERNAL_ERROR: 未预期的服务端异常，已记录日志（含 trace_id）。
        KG_TENANT_LEAK: 跨租户子图泄漏检测到（ADR-0003 §4，Sprint 5 批次 B）。active kg_version 内任一 Entity 节点 org_id 与当前 org_id 不一致——属数据质量事故伪装为正常结论，由 /agent/query 路由层转 403。
        KG_VERSION_NOT_ACTIVE: 请求显式指定了非 active 的 kg_version（writing / failed / superseded），按 ADR-0002 §3.2 一律拒绝，严禁静默降级到最新版。
        NOT_FOUND: 通用资源不存在（含未注册路由）。
        NOT_IMPLEMENTED: 基础设施不可用（Neo4j 图谱存储不可用（连不上 / 查询失败）/ LLM 未配置或装配失败）时返回 501，**不**表示「接口未实现」。
        TASK_INTERRUPTED: 进程重启导致在途任务被 TaskManager.recover() 回收置 failed（ADR-0001 §3.2）。
        UNAUTHORIZED: 缺少或无法解析认证态（Bearertoken / 开发态请求头）。
        UNSUPPORTED_MEDIA_TYPE: 上传文件 MIME 不在白名单（M1 验收 3）。
        VALIDATION_ERROR: 请求体 / 查询参数 / 路径参数未通过 Pydantic 校验。
      x-enum-sources:
        DOCUMENT_NOT_FOUND: M1 §5.4
        ENTITY_NOT_FOUND: Sprint 5 批次 C / plans §4.2 批次 C
        FILE_TOO_LARGE: M1 §3 验收 2
        FORBIDDEN: ADR-0003 §3.3 / M5 §3 验收 1
        HTTP_ERROR: CODEBUDDY.md 错误响应规范
        INTERNAL_ERROR: CODEBUDDY.md 错误响应规范
        KG_TENANT_LEAK: ADR-0003 §4 / Sprint 5 批次 B
        KG_VERSION_NOT_ACTIVE: ADR-0002 §3.2 / M3 §4.1
        NOT_FOUND: CODEBUDDY.md 错误响应规范
        NOT_IMPLEMENTED: backend/CODEBUDDY.md §1.1 故障语义边界 / ADR-0002 §3.2
        TASK_INTERRUPTED: ADR-0001 §3.2
        UNAUTHORIZED: M5 §3 验收 1
        UNSUPPORTED_MEDIA_TYPE: M1 §3 验收 3
        VALIDATION_ERROR: CODEBUDDY.md 错误响应规范
    ErrorResponse:
      description: '统一错误响应体（**所有** 4xx / 5xx 均使用本结构）。


        来源：`CODEBUDDY.md`「错误响应规范」。'
      example:
        code: KG_VERSION_NOT_ACTIVE
        detail:
          kg_version: 20260320T1430Z-01H9X9
          version_status: writing
        message: Requested kg_version is not active
        trace_id: 5f2c1b7e-9d4a-4c1e-8f3b-6a0d2e5c7b91
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
          description: 业务错误码（HTTP 状态码与业务错误码分离）
          examples:
          - VALIDATION_ERROR
          - FORBIDDEN
          - KG_VERSION_NOT_ACTIVE
        detail:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          description: 结构化补充信息；敏感字段（文件名原文 / error_detail / pii_flags）严禁出现在此
          title: Detail
        message:
          description: 面向调用方的简短英文摘要，便于日志检索
          title: Message
          type: string
        trace_id:
          description: 与响应头 X-Trace-Id 一致，贯穿 M1–M5 全链路
          title: Trace Id
          type: string
      required:
      - code
      - message
      - trace_id
      title: ErrorResponse
      type: object
    GraphEdge:
      properties:
        id:
          title: Id
          type: string
        properties:
          additionalProperties: true
          description: 关系属性（如 `share_pct` / `since` / `amount` / `role`），不含证据原文
          title: Properties
          type: object
        source:
          description: 起点节点 id
          title: Source
          type: string
        target:
          description: 终点节点 id
          title: Target
          type: string
        type:
          description: 关系类型（M2 §4.2），含桥梁抽取的实体↔实体类关系
          enum:
          - HAS_CHUNK
          - MENTIONS
          - SUPPORTED_BY
          - AFFILIATED_WITH
          - SUPPLIES_TO
          - PARTY_TO
          - HAS_FINANCIAL_INDICATOR
          - OPERATES_SEGMENT
          - RELATED
          title: Type
          type: string
      required:
      - id
      - type
      - source
      - target
      title: GraphEdge
      type: object
    GraphNode:
      description: '图谱节点。


        **刻意不含** `pii_flags`：M2 §5.3 明确禁止该字段外泄。'
      properties:
        canonical_name:
          anyOf:
          - type: string
          - type: 'null'
          description: '`:Entity.canonical_name`（消解后的标准名）'
          title: Canonical Name
        confidence:
          anyOf:
          - maximum: 1.0
            minimum: 0.0
            type: number
          - type: 'null'
          description: 抽取置信度
          title: Confidence
        entity_type:
          anyOf:
          - type: string
          - type: 'null'
          description: '`:Entity.type`（公司 / 自然人 / 法人 / 地址 / 证件号等）'
          title: Entity Type
        id:
          title: Id
          type: string
        kg_version:
          description: 节点所属图谱版本，必须等于响应的 `kg_version`
          title: Kg Version
          type: string
        label:
          enum:
          - Document
          - Chunk
          - Entity
          - Evidence
          title: Label
          type: string
      required:
      - id
      - label
      - kg_version
      title: GraphNode
      type: object
...
    UploadResponse:
      description: '`POST /api/v1/documents/upload` 响应（M1 验收 1）。'
      example:
        status: pending
        task_id: 3f1a9c2e-7b45-4d8a-9e01-2c4f6a8b0d11
        trace_id: 5f2c1b7e-9d4a-4c1e-8f3b-6a0d2e5c7b91
      properties:
        status:
          const: pending
          description: 上传受理后恒为 `pending`（异步执行）
          title: Status
          type: string
        task_id:
          description: 任务 id（UUIDv4），与 `documents.id` 一致；M2 抽取任务复用同一 id 空间
          format: uuid
          title: Task Id
          type: string
        trace_id:
          title: Trace Id
          type: string
      required:
      - task_id
      - status
      - trace_id
      title: UploadResponse
      type: object
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: 认证态来源。Sprint 1 支持 `Bearer dev.<org_id>.<actor_id>`；Sprint 3 起由 M5 `POST /auth/login` 签发的 JWT 取代。
      scheme: bearer
      type: http
info:
  description: "GraphRAG-Agent 后端 API（契约 v1.0，Sprint 1）。\n\n## 租户隔离（ADR-0003，**所有接口强制**）\n- `org_id` **只能**来自认证态：`Authorization: Bearer <token>` 解析结果；\n  **严禁**从请求 body / query 读取。\n- 开发态兜底（`ALLOW_DEV_ORG_HEADER=true` 且非生产）可用 `X-Org-Id` / `X-Actor-Id` 请求头，\n  该兜底 **Sprint 3 必须移除**。\n- 跨租户访问一律 `403 FORBIDDEN`；`org_id` 过滤**必须**经 ORM / 统一注入，禁止裸写 SQL。\n- **例外**：`GET /api/v1/health` 不要求认证、不返回任何业务数据。\n\n## trace_id（M1 验收 7 / M5 §5.3）\n- 请求头 `X-Trace-Id` 可选透传（必须是合法 UUID），缺失则服务端生成 UUIDv4；\n- 响应头 `X-Trace-Id` 恒回显；错误体 `trace_id` 与之一致。\n\n## 统一错误响应\n所有 4xx / 5xx 均为 `{code, message, detail, trace_id}`，`code` 取自 `ErrorCode` 枚举。\n\n## 图谱版本一致性（ADR-0002）\n查询层只认 `kg_version.status = active`。请求显式指定非 active 版本时，\n一律返回 `409 KG_VERSION_NOT_ACTIVE`，**严禁静默降级**到最新版本。\n\n## 实现状态（Sprint 4）\n五个接口均已实装真实链路；`501 NOT_IMPLEMENTED` 不再表示「接口未实现」，而是\n**基础设施故障**（Neo4j 不可用 / LLM 未配置或装配失败）的统一出口。"
  title: GraphRAG-Agent Backend API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /api/v1/documents:
    get:
      description: '按 `org_id` 强制过滤（ADR-0003），跨租户资源**永不**出现在结果里。


        **查询参数**：

        - `q`：按 `filename_hash`（SHA-256 hex）前缀匹配。文件名原文**不**落库，演示场景下做「输入前缀」匹配即可。

        - `status`：可选 `pending / processing / completed / failed`。

        - `page`：1-based 页码，默认 1。

        - `page_size`：默认 10，**上限 100**（演示前端表格默认 10 条 / 页）。


        **返回字段**：`filename` 字段展示的是 `filename_hash`（前端表格沿用 mock口径展示 hash，**不**伪造文件名原文，符合 M5 §4.5）；`entity_count` 仅当文档已参与建图（`kg_version_id IS NOT NULL`）时有值，否则为 `null`。


        **一致性**：列表按 `created_at DESC` 排序（最新上传在前），PG 唯一真源；Neo4j 不可用时本接口仍可正常返回——不依赖图谱存储。'
      operationId: listDocuments
      parameters:
      - description: 按 `filename_hash` 前缀匹配（SHA-256 hex）。示例：`q=e3b0` 匹配 hash 以 `e3b0` 开头的所有文档
        in: query
        name: q
        required: false
        schema:
          anyOf:
          - maxLength: 64
            minLength: 1
            type: string
          - type: 'null'
          description: 按 `filename_hash` 前缀匹配（SHA-256 hex）。示例：`q=e3b0` 匹配 hash 以 `e3b0` 开头的所有文档
          title: Q
      - description: 按状态过滤；M1 H1 状态机取值之一
        in: query
        name: status
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 按状态过滤；M1 H1 状态机取值之一
          title: Status
      - description: 1-based 页码
        in: query
        name: page
        required: false
        schema:
          default: 1
          description: 1-based 页码
          minimum: 1
          title: Page
          type: integer
      - description: 每页条目数（默认 10，上限 100）
        in: query
        name: page_size
        required: false
        schema:
          default: 10
          description: 每页条目数（默认 10，上限 100）
          maximum: 100
          minimum: 1
          title: Page Size
          type: integer
      - description: 【仅开发态兜底】租户 id。仅当 ALLOW_DEV_ORG_HEADER=true 且非生产环境时生效；Sprint 3 接入 M5 登录后必须移除（ADR-0003 §3.3：org_id 严禁来自 body / query）。
        in: header
        name: X-Org-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 【仅开发态兜底】租户 id。仅当 ALLOW_DEV_ORG_HEADER=true 且非生产环境时生效；Sprint 3 接入 M5 登录后必须移除（ADR-0003 §3.3：org_id 严禁来自 body / query）。
          title: X-Org-Id
      - description: 【仅开发态兜底】操作者 id，缺省取 DEFAULT_ACTOR_ID；Sprint 3 起由认证态提供。
        in: header
        name: X-Actor-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 【仅开发态兜底】操作者 id，缺省取 DEFAULT_ACTOR_ID；Sprint 3 起由认证态提供。
          title: X-Actor-Id
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 请求校验失败（`VALIDATION_ERROR`），`detail.errors` 给出字段级原因
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 缺少或无法解析认证态（`UNAUTHORIZED`）
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 跨租户访问被拒（`FORBIDDEN`，ADR-0003 §3.3 / M5 §3 验收 1）
      security:
      - bearerAuth: []
      summary: 列出当前租户的文档（批次 C）
      tags:
      - documents
  /api/v1/documents/upload:
    post:
      description: '**同步校验 + 立即返回**，不在请求内做任何解析：

        1. MIME 白名单校验 → 失败 415；

        2. 大小上限校验（默认 100MB）→ 失败 413；

        3. 落 `documents` 记录（`status = pending`），注册 `document.parse` 异步执行体，并返回 `task_id`。


        **解析链路**：执行体真实推进 `pending → processing → completed / failed`（M1 硬约束 H1），第三方 IO 异常按 H8 指数退避重试（≤ 3 次）；进程重启时由启动回收把在途任务置 `failed` + `error_code = TASK_INTERRUPTED`（ADR-0001 §3.2）。


        上传响应的 `status` 恒为 `pending`，后续进度请轮询 `GET /documents/{id}/status`。


        **当前局限**：状态机与错误落库为真实链路；MinerU 结构化解析与 LangExtract 实体关系抽取尚未接入，执行体当前返回空结果（后续版本补齐，见 v1.1.0 待办）。


        文件名以 SHA-256 落库（`filename_hash`），日志与响应均不含原文。'
      operationId: uploadDocument
      parameters:
      - description: 【仅开发态兜底】租户 id。仅当 ALLOW_DEV_ORG_HEADER=true 且非生产环境时生效；Sprint 3 接入 M5 登录后必须移除（ADR-0003 §3.3：org_id 严禁来自 body / query）。
        in: header
        name: X-Org-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 【仅开发态兜底】租户 id。仅当 ALLOW_DEV_ORG_HEADER=true 且非生产环境时生效；Sprint 3 接入 M5 登录后必须移除（ADR-0003 §3.3：org_id 严禁来自 body / query）。
          title: X-Org-Id
      - description: 【仅开发态兜底】操作者 id，缺省取 DEFAULT_ACTOR_ID；Sprint 3 起由认证态提供。
        in: header
        name: X-Actor-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 【仅开发态兜底】操作者 id，缺省取 DEFAULT_ACTOR_ID；Sprint 3 起由认证态提供。
          title: X-Actor-Id
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_uploadDocument'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 缺少或无法解析认证态（`UNAUTHORIZED`）
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 跨租户访问被拒（`FORBIDDEN`，ADR-0003 §3.3 / M5 §3 验收 1）
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 上传文件超过 `MAX_UPLOAD_SIZE_MB`（`FILE_TOO_LARGE`，M1 验收 2）
        '415':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: MIME 不在白名单（`UNSUPPORTED_MEDIA_TYPE`，M1 验收 3）
      security:
      - bearerAuth: []
      summary: 上传文档（异步受理）
      tags:
      - documents
  /api/v1/documents/{document_id}/graph:
    get:
      description: '返回该文档在 Neo4j 中的子图（`nodes` + `edges`），供前端力导向图渲染。


        **一致性（ADR-0002 §3.2）**：只返回 `status = active` 的 `kg_version`；该文档不存在 active 版本时返回 **409** `KG_VERSION_NOT_ACTIVE`，**绝不静默降级**到其他版本；响应中的 `version_status` 恒为 `active`。


        规模上限对齐 M3 §3 验收 1：单次节点数 ≤ 500，超限 `truncated = true`。


        **实现状态**：已实装——由 `GraphService.fetch_document_subgraph`（`app/services/graphs.py`）按 `kg_version` 查询 Neo4j 子图，并映射为 `DocumentGraphResponse`（`nodes` / `edges` / `truncated` / `version_status`）。


        **501 `NOT_IMPLEMENTED` 的真实语义**：Neo4j 不可用（连接失败 / 查询超时 / 凭据错误）属**基础设施故障**，此时返回 501——**不**表示「接口未实现」。'
      operationId: getDocumentGraph
      parameters:
      - in: path
        name: document_id
        required: true
        schema:
          format: uuid
          title: Document Id
          type: string
      - description: 【仅开发态兜底】租户 id。仅当 ALLOW_DEV_ORG_HEADER=true 且非生产环境时生效；Sprint 3 接入 M5 登录后必须移除（ADR-0003 §3.3：org_id 严禁来自 body / query）。
        in: header
        name: X-Org-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 【仅开发态兜底】租户 id。仅当 ALLOW_DEV_ORG_HEADER=true 且非生产环境时生效；Sprint 3 接入 M5 登录后必须移除（ADR-0003 §3.3：org_id 严禁来自 body / query）。
          title: X-Org-Id
      - description: 【仅开发态兜底】操作者 id，缺省取 DEFAULT_ACTOR_ID；Sprint 3 起由认证态提供。
        in: header
        name: X-Actor-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 【仅开发态兜底】操作者 id，缺省取 DEFAULT_ACTOR_ID；Sprint 3 起由认证态提供。
          title: X-Actor-Id
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentGraphResponse'
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 缺少或无法解析认证态（`UNAUTHORIZED`）
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 跨租户访问被拒（`FORBIDDEN`，ADR-0003 §3.3 / M5 §3 验收 1）
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 文档不存在（`DOCUMENT_NOT_FOUND`）。**跨租户访问返回 403 而非 404**（ADR-0003 / M5 §3 验收 1）
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 指定的 `kg_version` 非 active（`KG_VERSION_NOT_ACTIVE`）。**严禁静默降级**到最新 active 版本（ADR-0002 §3.2）
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 基础设施不可用（Neo4j 连接失败 / 查询超时，或 LLM 未配置、装配失败）时返回 501（`NOT_IMPLEMENTED`）
      security:
      - bearerAuth: []
      summary: 获取文档图谱子图
      tags:
      - documents
  /api/v1/documents/{document_id}/status:
    get:
      description: '状态机严格限定为 `pending → processing → completed / failed`（M1 硬约束 H1）。


        轮询建议：首次 2s，3 次后降为 10s。


        **跨租户访问返回 403**（`FORBIDDEN`），不返回 404——按 M5 §3 验收 1 的显式要求。'
      operationId: getDocumentStatus
      parameters:
      - in: path
        name: document_id
        required: true
        schema:
          format: uuid
          title: Document Id
          type: string
      - description: 【仅开发态兜底】租户 id。仅当 ALLOW_DEV_ORG_HEADER=true 且非生产环境时生效；Sprint 3 接入 M5 登录后必须移除（ADR-0003 §3.3：org_id 严禁来自 body / query）。
        in: header
        name: X-Org-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 【仅开发态兜底】租户 id。仅当 ALLOW_DEV_ORG_HEADER=true 且非生产环境时生效；Sprint 3 接入 M5 登录后必须移除（ADR-0003 §3.3：org_id 严禁来自 body / query）。
          title: X-Org-Id
      - description: 【仅开发态兜底】操作者 id，缺省取 DEFAULT_ACTOR_ID；Sprint 3 起由认证态提供。
        in: header
        name: X-Actor-Id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 【仅开发态兜底】操作者 id，缺省取 DEFAULT_ACTOR_ID；Sprint 3 起由认证态提供。
          title: X-Actor-Id
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentStatusResponse'
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 缺少或无法解析认证态（`UNAUTHORIZED`）
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 跨租户访问被拒（`FORBIDDEN`，ADR-0003 §3.3 / M5 §3 验收 1）
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 文档不存在（`DOCUMENT_NOT_FOUND`）。**跨租户访问返回 403 而非 404**（ADR-0003 / M5 §3 验收 1）
      security:
      - bearerAuth: []
      summary: 查询文档解析状态
      tags:
      - documents
...
servers:
- description: 同源部署（开发默认；生产由网关注入实际域名）
  url: /
...